+++ etc/block-components/ui-examples/ui/src/app.ts
@@ -1,6 +1,6 @@
 import { platforma } from '@milaboratories/milaboratories.ui-examples.model';
-import { defineApp } from '@platforma-sdk/ui-vue';
-import { computed, reactive } from 'vue';
+import { animate, defineApp, makeEaseInOut, makeEaseOut } from '@platforma-sdk/ui-vue';
+import { computed, reactive, ref } from 'vue';
 import LogViewPage from './pages/LogViewPage.vue';
 import ModalsPage from './pages/ModalsPage.vue';
 import InjectEnvPage from './pages/InjectEnvPage.vue';
@@ -15,6 +15,8 @@ import PlAgDataTablePage from './pages/PlAgDataTablePage.vue';
 import IconsPage from './pages/IconsPage.vue';
 import PlTextFieldPage from './pages/PlTextFieldPage.vue';
 import PlTabsPage from './pages/PlTabsPage.vue';
+import DraftsPage from './pages/DraftsPage.vue';
+import LayoutPage from './pages/LayoutPage.vue';
 
 export const sdkPlugin = defineApp(platforma, (base) => {
   // Additional data
@@ -28,25 +30,50 @@ export const sdkPlugin = defineApp(platforma, (base) => {
 
   const argsAsJson = computed(() => JSON.stringify(base.args));
 
+  const progressRef = ref<boolean | number>();
+
+  function showLoader(duration: number) {
+    progressRef.value = true;
+    setTimeout(() => (progressRef.value = false), duration);
+  }
+
+  function showProgress(duration: number) {
+    progressRef.value = 0;
+    animate({
+      duration,
+      timing: makeEaseOut((t) => t),
+      draw: (progress) => {
+        progressRef.value = progress;
+      }
+    });
+  }
+
   return {
     data,
     incrementCounter,
     argsAsJson,
+    showInfiniteProgress: showLoader,
+    showProgress,
+    progress: () => {
+      return progressRef.value;
+    },
     routes: {
       '/': IconsPage,
-      '/log-view': LogViewPage,
-      '/modals': ModalsPage,
-      '/inject-env': InjectEnvPage,
-      '/dropdowns': DropdownsPage,
-      '/use-watch-fetch': UseWatchFetchPage,
-      '/form-components': FormComponentsPage,
-      '/typography': TypographyPage,
-      '/ag-grid-vue': AgGridVuePage,
+      '/layout': () => LayoutPage,
+      '/log-view': () => LogViewPage,
+      '/modals': () => ModalsPage,
+      '/inject-env': () => InjectEnvPage,
+      '/dropdowns': () => DropdownsPage,
+      '/use-watch-fetch': () => UseWatchFetchPage,
+      '/form-components': () => FormComponentsPage,
+      '/typography': () => TypographyPage,
+      '/ag-grid-vue': () => AgGridVuePage,
       '/pl-ag-data-table': () => PlAgDataTablePage,
-      '/select-files': SelectFilesPage,
+      '/select-files': () => SelectFilesPage,
       '/errors': () => ErrorsPage,
       '/text-fields': () => PlTextFieldPage,
-      '/tabs': () => PlTabsPage
+      '/tabs': () => PlTabsPage,
+      '/drafts': () => DraftsPage
     }
   };
 });