+++ lib/ui/uikit/src/components/PlSectionSeparator/index.ts
@@ -1,0 +1,1 @@
+export { default as PlSectionSeparator } from './PlSectionSeparator.vue';