+++ sdk/ui-vue/src/components/BlockLayout.vue
@@ -6,6 +6,7 @@ import { useSdkPlugin } from '../defineApp';
 import NotFound from './NotFound.vue';
 import LoaderPage from './LoaderPage.vue';
 import { PlAppErrorNotificationAlert } from './PlAppErrorNotificationAlert';
+import BlockLoader from './BlockLoader.vue';
 
 const sdk = useSdkPlugin();
 
@@ -35,10 +36,13 @@ const app = computed(() => (sdk.loaded ? sdk.useApp() : undefined));
 const errors = computed(() => (app.value ? app.value.model.outputErrors : {}));
 
 const showErrorsNotification = computed(() => app.value?.showErrorsNotification ?? true);
+
+const progress = computed(() => app.value?.progress?.());
 </script>
 
 <template>
   <div class="block block__layout">
+    <BlockLoader :value="progress" />
     <div v-if="sdk.error">{{ sdk.error }}</div>
     <LoaderPage v-else-if="!sdk.loaded">Loading...</LoaderPage>
     <component :is="CurrentView" v-else-if="CurrentView" :key="href" />