+++ sdk/ui-vue/src/components/PlAgTextAndButtonCell/index.ts
@@ -1,0 +1,1 @@
+export { default as PlAgTextAndButtonCell } from './PlAgTextAndButtonCell.vue';