+++ sdk/ui-vue/src/defineApp.ts
@@ -3,7 +3,7 @@ import { type BlockOutputsBase, type Platforma } from '@platforma-sdk/model';
 import type { Component, Reactive } from 'vue';
 import { inject, markRaw, reactive } from 'vue';
 import { createApp, type BaseApp } from './internal/createApp';
-import type { BaseSettings, Routes } from './types';
+import type { AppSettings, ExtendSettings, Routes } from './types';
 import { activateAgGrid } from './aggrid';
 
 const pluginKey = Symbol('sdk-vue');
@@ -18,12 +18,13 @@ export function defineApp<
   Outputs extends BlockOutputsBase = BlockOutputsBase,
   UiState = unknown,
   Href extends `/${string}` = `/${string}`,
-  Local extends BaseSettings<Href> = BaseSettings<Href>,
+  Extend extends ExtendSettings<Href> = ExtendSettings<Href>,
 >(
   platforma: Platforma<Args, Outputs, UiState, Href>,
-  extendApp: (app: BaseApp<Args, Outputs, UiState, Href>) => Local,
-): SdkPlugin<Args, Outputs, UiState, Href, Local> {
-  let app: undefined | App<Args, Outputs, UiState, Href, Local> = undefined;
+  extendApp: (app: BaseApp<Args, Outputs, UiState, Href>) => Extend,
+  settings: AppSettings = {},
+): SdkPlugin<Args, Outputs, UiState, Href, Extend> {
+  let app: undefined | App<Args, Outputs, UiState, Href, Extend> = undefined;
 
   activateAgGrid();
 
@@ -32,7 +33,7 @@ export function defineApp<
       .loadBlockState()
       .then((state) => {
         plugin.loaded = true;
-        const baseApp = createApp<Args, Outputs, UiState, Href>(state, platforma);
+        const baseApp = createApp<Args, Outputs, UiState, Href>(state, platforma, settings);
 
         const localState = extendApp(baseApp);
 
@@ -48,7 +49,7 @@ export function defineApp<
           getRoute(href: Href): Component | undefined {
             return routes[href];
           },
-        } as unknown as App<Args, Outputs, UiState, Href, Local>);
+        } as unknown as App<Args, Outputs, UiState, Href, Extend>);
       })
       .catch((err) => {
         console.error('load initial state error', err);
@@ -61,7 +62,7 @@ export function defineApp<
     error: undefined as unknown,
     // Href to get typed query parameters for a specific route
     useApp<PageHref extends Href = Href>() {
-      return notEmpty(app, 'App is not loaded') as App<Args, Outputs, UiState, PageHref, Local>;
+      return notEmpty(app, 'App is not loaded') as App<Args, Outputs, UiState, PageHref, Extend>;
     },
     // @todo type portability issue with Vue
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -79,7 +80,7 @@ export type App<
   Outputs extends BlockOutputsBase = BlockOutputsBase,
   UiState = unknown,
   Href extends `/${string}` = `/${string}`,
-  Local extends BaseSettings<Href> = BaseSettings<Href>,
+  Local extends ExtendSettings<Href> = ExtendSettings<Href>,
 > = BaseApp<Args, Outputs, UiState, Href> & Reactive<Omit<Local, 'routes'>> & { getRoute(href: Href): Component | undefined };
 
 export type SdkPlugin<
@@ -87,7 +88,7 @@ export type SdkPlugin<
   Outputs extends BlockOutputsBase = BlockOutputsBase,
   UiState = unknown,
   Href extends `/${string}` = `/${string}`,
-  Local extends BaseSettings<Href> = BaseSettings<Href>,
+  Local extends ExtendSettings<Href> = ExtendSettings<Href>,
 > = {
   loaded: boolean;
   error: unknown;