+++ sdk/ui-vue/src/types.ts
@@ -51,8 +51,36 @@ export type RouteParams<Href extends `/${string}` = `/${string}`> = {
   [P in Href as ParsePathnamePart<P>]: ParseQuery<P>;
 };
 
-export type BaseSettings<Href extends `/${string}` = `/${string}`> = {
+export type AppSettings = {
+  /**
+   * Deep patch `model.` This parameter affects how you listen to model changes; a deep watcher is required if you use the `source` function.
+   */
+  deepPatchModel?: boolean;
+  /**
+   * Enables some debug logs
+   */
+  debug?: boolean;
+};
+
+export type ExtendSettings<Href extends `/${string}` = `/${string}`> = {
+  /**
+   * Displays a loader on top of the block.
+   * - If `true`: Shows an infinite loader.
+   * - If a number (0 <= n < 1): Displays a progress bar representing completion percentage.
+   * - If a number (n > 1): Hides the progress bar.
+   * - If `undefined`: No loader is displayed.
+   */
+  progress?: () => number | boolean | undefined;
+  /**
+   * Enables or disables a notification box for error messages.
+   * - If `true`: A notification box appears when there are errors in the outputs.
+   * - If `false`: No notification box is shown.
+   */
   showErrorsNotification?: boolean;
+  /**
+   * Maps application routes to their respective components.
+   * TODO: Consider moving route initialization logic to a dedicated method.
+   */
   routes: Routes<Href>;
 };