+++ etc/blocks/enter-numbers/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @milaboratories/milaboratories.test-enter-numbers
 
+## 1.0.104
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.104
+
+## 1.0.103
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.103
+
+## 1.0.102
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.102
+
+## 1.0.101
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.101
+
+## 1.0.100
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.100
+
+## 1.0.99
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.99
+
+## 1.0.98
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.98
+
+## 1.0.97
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.97
+
+## 1.0.96
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.96
+
+## 1.0.95
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.95
+
 ## 1.0.94
 
 ### Patch Changes