+++ etc/blocks/model-test/block/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @milaboratories/milaboratories.test-block-model
 
+## 1.0.23
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.23
+
+## 1.0.22
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.22
+
+## 1.0.21
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.21
+
+## 1.0.20
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.20
+
+## 1.0.19
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.19
+
+## 1.0.18
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.18
+
+## 1.0.17
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.17
+
+## 1.0.16
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.16
+
+## 1.0.15
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.15
+
+## 1.0.14
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.14
+
 ## 1.0.13
 
 ### Patch Changes