+++ etc/blocks/model-test/ui/CHANGELOG.md
@@ -1,5 +1,73 @@
 # @milaboratories/milaboratories.test-block-model.ui
 
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies [d60c614]
+  - @platforma-sdk/ui-vue@1.11.8
+
+## 1.0.22
+
+### Patch Changes
+
+- Updated dependencies [12c72ef]
+  - @platforma-sdk/ui-vue@1.11.7
+
+## 1.0.21
+
+### Patch Changes
+
+- Updated dependencies [9cd0c99]
+  - @platforma-sdk/ui-vue@1.11.6
+
+## 1.0.20
+
+### Patch Changes
+
+- @platforma-sdk/ui-vue@1.11.5
+
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [5086e90]
+  - @platforma-sdk/ui-vue@1.11.4
+
+## 1.0.18
+
+### Patch Changes
+
+- @platforma-sdk/ui-vue@1.11.3
+
+## 1.0.17
+
+### Patch Changes
+
+- Updated dependencies [4783725]
+  - @platforma-sdk/ui-vue@1.11.2
+
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [961b8b1]
+  - @platforma-sdk/ui-vue@1.11.1
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [680e8b8]
+  - @platforma-sdk/ui-vue@1.11.0
+
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [166c161]
+  - @platforma-sdk/ui-vue@1.10.18
+
 ## 1.0.13
 
 ### Patch Changes