+++ etc/uikit-playground/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @milaboratories/uikit-playground
 
+## 0.0.58
+
+### Patch Changes
+
+- Updated dependencies [9a1140c]
+  - @milaboratories/uikit@2.2.11
+
+## 0.0.57
+
+### Patch Changes
+
+- Updated dependencies [ec5be92]
+  - @milaboratories/uikit@2.2.10
+
+## 0.0.56
+
+### Patch Changes
+
+- 166c161: remove global hidden webkit-srollbar
+- Updated dependencies [166c161]
+  - @milaboratories/uikit@2.2.9
+
 ## 0.0.55
 
 ### Patch Changes